requirements.txt
# Dashboard L'Acqua Azzurra - Dependências de Produção
# Python 3.13+

# Framework principal
dash==3.3.0
dash-bootstrap-components==2.0.4
plotly==6.5.0

# Processamento de dados
pandas==2.3.3
numpy==2.3.5
openpyxl==3.1.5

# Banco de dados PostgreSQL
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1

# Servidor WSGI para produção
gunicorn==23.0.0

# Dependências do Flask/Dash
Flask==3.1.2
Werkzeug==3.1.4
click==8.3.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
blinker==1.9.0

# Utilidades
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
requests==2.32.5
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
urllib3==2.5.0
packaging==25.0
typing_extensions==4.15.0
importlib_metadata==8.7.0
zipp==3.23.0
six==1.17.0
